requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asyncer==0.0.8
asyncpg==0.30.0
attrs==25.4.0
backoff==2.2.1
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
dspy==3.0.3
dspy-ai==3.0.3
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.116.1
fastuuid==0.13.5
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
fuzzywuzzy==0.18.0
gepa==0.0.7
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
inflect==7.5.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json_repair==0.52.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Levenshtein==0.27.1
litellm==1.77.7
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
motor==3.7.1
multidict==6.7.0
numpy==2.3.2
openai==1.107.0
openpyxl==3.1.5
optuna==4.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
propcache==0.4.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
sqlacodegen==3.1.1
SQLAlchemy==2.0.41
starlette==0.47.3
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
mangum
sentence-transformers>=2.2.0
